+++ module/actor/actor-cm.mjs
@@ -753,3 +753,129 @@ export const fatigueAttMenuOptions = (actor,token) => [
     }
   }
 ];
+
+//Allegiance Item Context Menu Options 
+export const allegianceMenuOptions = (actor,token) => [
+  {
+    name: game.i18n.localize("BRP.allegiance"),
+    icon: "",
+    condition: () => true,
+    callback: (el) => {}
+  },
+  {
+    name: game.i18n.localize("BRP.card.NO"),
+    icon: '<i class="fas fa-dice-d20"></i>',
+    condition: () => true,
+    callback: (el) => {
+      BRPCheck._trigger({
+        rollType: 'AL',
+        cardType: 'NO',
+        skillId: el[0].dataset.itemId,
+        actor,
+        token,}
+      );
+    }
+  },  
+  {
+    name: game.i18n.localize("BRP.improve")+ " (" + game.settings.get('brp','xpFormula') + ")",
+    icon: '<i class="fas fa-dice-d6"></i>',
+    condition: (el) => (game.settings.get('brp','development') && el[0].dataset.xp === 'true'),
+    callback: (el) => {
+      BRPCharDev.onXPGainSingle(el[0].dataset.itemId,actor,token,"formula");
+    }
+  },    
+  {
+    name: game.i18n.localize("BRP.improve")+ " (" + game.settings.get('brp','xpFixed') + ")",
+    icon: '<i class="fas fa-dice-three"></i>',
+    condition: (el) => (game.settings.get('brp','development') && el[0].dataset.xp === 'true'),
+    callback: (el) => {
+      BRPCharDev.onXPGainSingle(el[0].dataset.itemId,actor,token,"fixed");
+    }
+  }, 
+  {
+    name: game.i18n.localize("BRP.view"),
+    icon: '<i class="fas fa-magnifying-glass"></i>',
+    condition: () => true,
+    callback: (el) => {
+      const itemId = BRPUtilities.triggerEdit(el, actor, "itemId");
+    }
+  },
+  {
+    name: game.i18n.localize("BRP.delete"),
+    icon: '<i class="fas fa-trash"></i>',
+    condition: () => true,
+    callback: (el) => {
+      const itemId = BRPUtilities.triggerDelete(el, actor, "itemId");
+    }
+  }
+]; 
+
+//Passion Item Context Menu Options 
+export const passionMenuOptions = (actor,token) => [
+  {
+    name: game.i18n.localize("BRP.allegiance"),
+    icon: "",
+    condition: () => true,
+    callback: (el) => {}
+  },
+  {
+    name: game.i18n.localize("BRP.card.NO"),
+    icon: '<i class="fas fa-dice-d20"></i>',
+    condition: () => true,
+    callback: (el) => {
+      BRPCheck._trigger({
+        rollType: 'PA',
+        cardType: 'NO',
+        skillId: el[0].dataset.itemId,
+        actor,
+        token,}
+      );
+    }
+  },  
+  {
+    name: game.i18n.localize("BRP.card.OP"),
+    icon: '<i class="fas fa-handshake-simple"></i>',
+    condition: () => true,
+    callback: (el) => {
+      BRPCheck._trigger({
+        rollType: 'PA',
+        cardType: 'OP',
+        skillId: el[0].dataset.itemId,
+        actor,
+        token,}
+      );
+    }
+  },  
+  {
+    name: game.i18n.localize("BRP.improve")+ " (" + game.settings.get('brp','xpFormula') + ")",
+    icon: '<i class="fas fa-dice-d6"></i>',
+    condition: (el) => (game.settings.get('brp','development') && el[0].dataset.xp === 'true'),
+    callback: (el) => {
+      BRPCharDev.onXPGainSingle(el[0].dataset.itemId,actor,token,"formula");
+    }
+  },    
+  {
+    name: game.i18n.localize("BRP.improve")+ " (" + game.settings.get('brp','xpFixed') + ")",
+    icon: '<i class="fas fa-dice-three"></i>',
+    condition: (el) => (game.settings.get('brp','development') && el[0].dataset.xp === 'true'),
+    callback: (el) => {
+      BRPCharDev.onXPGainSingle(el[0].dataset.itemId,actor,token,"fixed");
+    }
+  }, 
+  {
+    name: game.i18n.localize("BRP.view"),
+    icon: '<i class="fas fa-magnifying-glass"></i>',
+    condition: () => true,
+    callback: (el) => {
+      const itemId = BRPUtilities.triggerEdit(el, actor, "itemId");
+    }
+  },
+  {
+    name: game.i18n.localize("BRP.delete"),
+    icon: '<i class="fas fa-trash"></i>',
+    condition: () => true,
+    callback: (el) => {
+      const itemId = BRPUtilities.triggerDelete(el, actor, "itemId");
+    }
+  }
+];